Commit perché se no non mi funziona il test sul branch di Matteo

+++ src/main/java/it/unitn/webprog2018/ueb/shoppinglist/servlets/admin/NewProductsCategoryServlet.java
@@ -7,13 +7,11 @@ package it.unitn.webprog2018.ueb.shoppinglist.servlets.admin;
 
 import it.unitn.webprog2018.ueb.shoppinglist.dao.DAOFactory;
 import it.unitn.webprog2018.ueb.shoppinglist.dao.exceptions.DaoException;
-import it.unitn.webprog2018.ueb.shoppinglist.dao.exceptions.RecordNotFoundDaoException;
 import it.unitn.webprog2018.ueb.shoppinglist.dao.interfaces.ProductsCategoryDAO;
 import it.unitn.webprog2018.ueb.shoppinglist.entities.ProductsCategory;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.PrintWriter;
 import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
@@ -69,15 +67,19 @@ public class NewProductsCategoryServlet extends HttpServlet {
 	@Override
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
+		String context = getServletContext().getContextPath();
+		if (!context.endsWith("/")) {
+			context += "/";
+		}
+
 		ProductsCategoryDAO productsCategoryDAO = ((DAOFactory) getServletContext().getAttribute("daoFactory")).getProductsCategoryDAO();
 		// parametri stringhe
 		String name = request.getParameter("name");
 		String description = request.getParameter("description");
 		//parametri file
 		String logoURI = "";
-		File fileLogo = null;
-		String logoFileName = "";
-		String logoFolder = getServletContext().getInitParameter("uploadFolder") + File.separator + "public" + File.separator + "productCategoryLogo" + File.separator;
+		String logoFolder = (String) getServletContext().getAttribute("categoryLogoFolder");
+		String uploadFolder = (String) getServletContext().getAttribute("uploadFolder");
 		Part logo = request.getPart("logo");
 		ProductsCategory productCategory = new ProductsCategory();
 		try {
@@ -88,33 +90,35 @@ public class NewProductsCategoryServlet extends HttpServlet {
 			if (productsCategoryDAO.addProductsCategory(productCategory)) {
 				//upload logo
 				if ((logo != null) && (logo.getSize() > 0)) {
-					logoFileName = Paths.get(logo.getSubmittedFileName()).getFileName().toString();
+					String logoFileName = Paths.get(logo.getSubmittedFileName()).getFileName().toString();
 					int ext = logoFileName.lastIndexOf(".");
 					int noExt = logoFileName.lastIndexOf(File.separator);
-					logoFileName = logoFolder + productCategory.getId() + (ext > noExt ? logoFileName.substring(ext) : "");
-					InputStream fileContentLogo = null;
+					logoFileName = logoFolder + productCategory.getHash() + (ext > noExt ? logoFileName.substring(ext) : "");
 					try {
-						ext = logoFileName.lastIndexOf(".");
-						noExt = logoFileName.lastIndexOf(File.separator);
-						fileContentLogo = logo.getInputStream();
-						fileLogo = new File(logoFileName);
-						Files.copy(fileContentLogo, fileLogo.toPath());
-						logoURI = File.separator + "uploads" + File.separator + "public" + File.separator + "productCategoryLogo"
-								+ File.separator + productCategory.getId() + (ext > noExt ? logoFileName.substring(ext) : "");
+						File fileLogo = new File(logoFileName);
+						if (fileLogo.exists()) {
+							fileLogo.delete();
+						}
+						Files.copy(logo.getInputStream(), fileLogo.toPath());
+						logoURI = context + logoFileName.substring(logoFileName.lastIndexOf(uploadFolder) + uploadFolder.length());
 
-					} catch (FileAlreadyExistsException ex) {
-						fileLogo.delete();
-						Files.copy(fileContentLogo, fileLogo.toPath());
-						logoURI = File.separator + "uploads" + File.separator + "public" + File.separator + "productCategoryLogo"
-								+ File.separator + productCategory.getId() + (ext > noExt ? logoFileName.substring(ext) : "");
-
+					} catch (IOException ex) {
+						// It is not a fatal error, we ask the user to try again
+						Logger.getLogger(NewProductsCategoryServlet.class.getName()).log(Level.WARNING, null, ex);
+						productCategory.setError("image", "Non è stato possibile salvare l'immagine, riprova più tardi o contatta un amministratore");
+						// allows to forward response with correct loading of accessory information from the database to be shown in the jsp.
+						doGet(request, response);
 					}
-					productCategory.setLogo(logoURI);
-					if (!productsCategoryDAO.updateProductsCategory(productCategory.getId(), productCategory)) {
-						response.sendError(500, "Qualcosa è andato storto. Non è stato possibile aggiornare logo");
+					if (!response.isCommitted()) {
+						productCategory.setLogo(logoURI);
+						if (!productsCategoryDAO.updateProductsCategory(productCategory.getId(), productCategory)) {
+							response.sendError(500, "Qualcosa è andato storto. Non è stato possibile aggiornare logo");
+						}
 					}
 				}
-				response.sendRedirect(getServletContext().getContextPath() + "/restricted/admin/ProductsCategory");
+				if (!response.isCommitted()) {
+					response.sendRedirect(getServletContext().getContextPath() + "/restricted/admin/ProductsCategory");
+				}
 			} else {
 				request.setAttribute("productCategory", productCategory);
 				request.getRequestDispatcher("/WEB-INF/views/admin/NewProductsCategory.jsp").forward(request, response);
